pyproject.toml
[project]
name = "stratlake-trade-engine"
version = "0.1.0"
description = "Research and backtesting engine built on curated marketlake data."
authors = [{ name = "Christopher Overton" }]
requires-python = ">=3.10"

dependencies = [
    "pandas",
    "matplotlib",
    "duckdb",
    "jsonschema",
    "pyarrow",
    "python-dotenv",
    "pyyaml",
    "scikit-learn",
    "xgboost",
    "lightgbm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint.per-file-ignores]
"docs/examples/**/*.py" = ["E402"]
"docs/examples/**/*.ipynb" = ["E402"]
